refactor: Moved files to their respective folders to keep smth like a clean architecture
