+++ library/src/main/java/com/yausername/youtubedl_android/data/local/streams/DownloadProgressCallback.kt
@@ -1,4 +1,4 @@
-package com.yausername.youtubedl_android
+package com.yausername.youtubedl_android.data.local.streams
 
 interface DownloadProgressCallback {
     fun onProgressUpdate(progress: Float, etaInSeconds: Long, line: String?)
