+++ library/src/main/java/com/yausername/youtubedl_android/data/local/streams/DownloadProgressCallbackImpl.kt
@@ -1,0 +1,7 @@
+package com.yausername.youtubedl_android.data.local.streams
+
+class DownloadProgressCallbackImpl : DownloadProgressCallback {
+    override fun onProgressUpdate(progress: Float, etaInSeconds: Long, line: String?) {
+        TODO("Not yet implemented")
+    }
+}
