+++ library/src/main/java/com/yausername/youtubedl_android/data/local/streams/StreamGobbler.kt
@@ -1,24 +1,27 @@
-package com.yausername.youtubedl_android
+package com.yausername.youtubedl_android.data.local.streams
 
 import android.util.Log
+import com.yausername.youtubedl_android.BuildConfig
 import java.io.IOException
 import java.io.InputStream
 import java.io.InputStreamReader
 import java.io.Reader
 import java.nio.charset.StandardCharsets
 
-internal class StreamGobbler(private val buffer: StringBuffer, private val stream: InputStream) :
+internal class StreamGobbler(private val buffer: StringBuilder, private val stream: InputStream) :
     Thread() {
+
     init {
         start()
     }
 
     override fun run() {
         try {
-            val `in`: Reader = InputStreamReader(stream, StandardCharsets.UTF_8)
-            var nextChar: Int
-            while (`in`.read().also { nextChar = it } != -1) {
-                buffer.append(nextChar.toChar())
+            InputStreamReader(stream, StandardCharsets.UTF_8).use { reader ->
+                var nextChar: Int
+                while (reader.read().also { nextChar = it } != -1) {
+                    buffer.append(nextChar.toChar())
+                }
             }
         } catch (e: IOException) {
             if (BuildConfig.DEBUG) Log.e(TAG, "failed to read stream", e)
