+++ library/src/main/java/com/yausername/youtubedl_android/data/local/streams/StreamProcessExtractor.kt
@@ -1,6 +1,7 @@
-package com.yausername.youtubedl_android
+package com.yausername.youtubedl_android.data.local.streams
 
 import android.util.Log
+import com.yausername.youtubedl_android.BuildConfig
 import java.io.IOException
 import java.io.InputStream
 import java.io.InputStreamReader
@@ -9,7 +10,7 @@ import java.nio.charset.StandardCharsets
 import java.util.regex.Pattern
 
 internal class StreamProcessExtractor(
-    private val buffer: StringBuffer,
+    private val buffer: StringBuilder,
     private val stream: InputStream,
     private val callback: ((Float, Long, String) -> Unit)?
 ) : Thread() {
@@ -25,18 +26,19 @@ internal class StreamProcessExtractor(
 
     override fun run() {
         try {
-            val input: Reader = InputStreamReader(stream, StandardCharsets.UTF_8)
-            val currentLine = StringBuilder()
-            var nextChar: Int
-            while (input.read().also { nextChar = it } != -1) {
-                buffer.append(nextChar.toChar())
-                if (nextChar == '\r'.code || nextChar == '\n'.code && callback != null) {
-                    val line = currentLine.toString()
-                    if (line.startsWith("[")) processOutputLine(line)
-                    currentLine.setLength(0)
-                    continue
+            InputStreamReader(stream, StandardCharsets.UTF_8).use { input ->
+                val currentLine = StringBuilder()
+                var nextChar: Int
+                while (input.read().also { nextChar = it } != -1) {
+                    buffer.append(nextChar.toChar())
+                    if (nextChar == '\r'.code || nextChar == '\n'.code && callback != null) {
+                        val line = currentLine.toString()
+                        if (line.startsWith("[")) processOutputLine(line)
+                        currentLine.setLength(0)
+                        continue
+                    }
+                    currentLine.append(nextChar.toChar())
                 }
-                currentLine.append(nextChar.toChar())
             }
         } catch (e: IOException) {
             if (BuildConfig.DEBUG) Log.e(TAG, "failed to read stream", e)
@@ -49,31 +51,29 @@ internal class StreamProcessExtractor(
 
     private fun getProgress(line: String): Float {
         val matcher = p.matcher(line)
-        if (matcher.find()) return matcher.group(GROUP_PERCENT).toFloat()
-            .also { progress = it } else {
-            val mAria2c = pAria2c.matcher(line)
-            if (mAria2c.find()) return mAria2c.group(1).toFloat().also { progress = it }
-        }
+        if (matcher.find()) return matcher.group(GROUP_PERCENT).toFloat().also { progress = it }
+        val mAria2c = pAria2c.matcher(line)
+        if (mAria2c.find()) return mAria2c.group(1).toFloat().also { progress = it }
         return progress
     }
 
     private fun getEta(line: String): Long {
         val matcher = p.matcher(line)
         if (matcher.find()) return convertToSeconds(
-            matcher.group(GROUP_MINUTES),
-            matcher.group(GROUP_SECONDS)
-        ).also { eta = it.toLong() }.toLong() else {
-            val mAria2c = pAria2c.matcher(line)
-            if (mAria2c.find()) return convertToSeconds(
-                mAria2c.group(3),
-                mAria2c.group(5)
-            ).also { eta = it.toLong() }.toLong()
-        }
+            minutes = matcher.group(GROUP_MINUTES),
+            seconds = matcher.group(GROUP_SECONDS)
+        ).also { eta = it.toLong() }.toLong()
+        val mAria2c = pAria2c.matcher(line)
+        if (mAria2c.find()) return convertToSeconds(
+            minutes = mAria2c.group(3),
+            seconds = mAria2c.group(5)
+        ).also { eta = it.toLong() }.toLong()
         return eta
     }
 
     private fun convertToSeconds(minutes: String?, seconds: String?): Int {
-        if (seconds == null) return 0 else if (minutes == null) return seconds.toInt()
+        if (seconds == null) return 0
+        if (minutes == null) return seconds.toInt()
         return minutes.toInt() * 60 + seconds.toInt()
     }
