+++ library/src/main/java/com/yausername/youtubedl_android/domain/model/VideoInfo.kt
@@ -1,4 +1,4 @@
-package com.yausername.youtubedl_android.mapper
+package com.yausername.youtubedl_android.domain.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
