+++ library/src/main/java/com/yausername/youtubedl_android/domain/model/updates/Uploader.kt
@@ -1,0 +1,25 @@
+package com.yausername.youtubedl_android.domain.model.updates
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Uploader(
+    val login: String = "",
+    val id: Long = 0,
+    val node_id: String = "",
+    val avatar_url: String = "",
+    val gravatar_id: String = "",
+    val url: String = "",
+    val html_url: String = "",
+    val followers_url: String = "",
+    val following_url: String = "",
+    val gists_url: String = "",
+    val starred_url: String = "",
+    val subscriptions_url: String = "",
+    val organizations_url: String = "",
+    val repos_url: String = "",
+    val events_url: String = "",
+    val received_events_url: String = "",
+    val type: String = "",
+    val site_admin: Boolean = false
+)
